rewrote component folder to typescript
