--- src/components/Navigation/SideDrawer/DrawerToggle/DrawerToggle.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import classes from './DrawerToggle.module.css';
-
-const drawerToggle = (props) => (
-    <div className={classes.DrawerToggle} onClick={props.clicked}>
-        <div></div>
-        <div></div>
-        <div></div>
-    </div>
-);
-
-export default drawerToggle;
