--- src/components/Navigation/SideDrawer/SideDrawer.js
@@ -1,29 +1,0 @@
-import React from 'react';
-
-import Logo from '../../Logo/Logo';
-import NavigationItems from '../NavigationItems/NavigationItems';
-import classes from './SideDrawer.module.css';
-import Backdrop from '../../UI/Backdrop/Backdrop';
-import Aux from '../../../hoc/_Aux/_Aux';
-
-const sideDrawer = ( props ) => {
-    let attachedClasses = [classes.SideDrawer, classes.Close];
-    if (props.open) {
-        attachedClasses = [classes.SideDrawer, classes.Open];
-    }
-    return (
-        <Aux>
-            <Backdrop show={props.open} clicked={props.closed}/>
-            <div className={attachedClasses.join(' ')}>
-                <div className={classes.Logo}>
-                    <Logo />
-                </div>
-                <nav>
-                    <NavigationItems isAuthentication={props.isAuthentication}/>
-                </nav>
-            </div>
-        </Aux>
-    );
-};
-
-export default sideDrawer;
